Teach #+TODO parser to understand (..) suffix (#184)
The in-buffer settings supported by both Org and organice include
modifying the TODO keywords, e.g.

    #+TODO: START INPROGRESS STALLED | FINISHED

or

    #+TYP_TODO: START INPROGRESS STALLED | FINISHED

Org supports adding a suffix in parentheses to any of these keywords
in order to configure fast-access keyboard shortcuts:

    https://orgmode.org/manual/Fast-access-to-TODO-states.html#Fast-access-to-TODO-states

or to track TODO state changes:

    https://orgmode.org/manual/Tracking-TODO-state-changes.html#Tracking-TODO-state-changes

For example:

    #+TODO: START(s) INPROGRESS(i!) STALLED(.@/!) | FINISHED(f@)

However, organice gets hopelessly confused by files which use this
syntax as described in #184.  So make the parser smarter, initially
simply by ignoring these suffixes so that organice will not change
behaviour, in a way which allows taking advantage of them later on.

Fixes #184.