Remove unit tests duplicated in parse_org.unit.test.js
There were a bunch of tests identically or near identically duplicated
between parse_org.unit.test.js and OrgFile.unit.test.js, so remove
them from one of the two files.  The former seems to be a more
specific and low-level set of tests; based on the file name presumably
it's the best place for testing parseOrg(), but probably less suitable
for tests which also exercise exportOrg().