doc: Use a more robust approach to enabling CORS pre-flight checks
Using the R=200 trick in mod_rewrite seems problematic, because it can
cause Apache to return a 200 OK response with a very unhelpful message
body including:

    The server encountered an internal error or misconfiguration and
    was unable to complete your request.

This SO answer suggests that the internal error is due to Apache not
having an ErrorDocument configured for 200 responses by default:

    https://stackoverflow.com/a/48529566/179332

Apparently a 204 code (No Content) is a more appropriate response:

    https://stackoverflow.com/questions/27703871/return-empty-response-from-apache/

and indeed my experiments confirm that changing the RewriteRule to use
R=204 eliminates this error.

However, this approach doesn't guarantee that the CORS headers are
included in the response.  A better approach is to revisit the
original motivation for this rewrite rule, which was presumably to
get around the fact that the browser doesn't send credentials
with the CORS preflight check, therefore the server would typically
give a 401 response.  So we can simply follow an alternative approach:

    https://serverfault.com/a/431718/79729

where we do not require authorization for OPTIONS requests.