Double performance of obtaining header and index for given header id
Many places in the code had adopted the following pattern:

  const header = headerWithId(headers, headerId);
  const headerIndex = indexOfHeaderWithId(headers, headerId);

However this is twice as slow as necessary, because headerWithId()
calls indexOfHeaderWithId(), and then throws away the result after
using it to return the header.

So instead build a single indexAndHeaderWithId() function which
obtains the index, does an O(1) lookup to obtain the header, then
returns an object containing both.

Partially addresses #321.