Handle a checkbox in an invalid state: throw in dev; ignore in prod
Following up on the discussion at: https://github.com/200ok-ch/organice/pull/431#issuecomment-680136909

If a checkboxState is not one of 'checked', 'unchecked', or 'partial',
the checkbox propagation now throws an exception to enable the developer fix
the erroneous state or extend the switch statement for the new state.

In a production deployment the code behaves as before.