Make Docker containers configurable
This change makes it possible to configure the application when
run using the official Docker image by setting the relevant
environment variables or mounting a `.env` file.

It does this by delaying building assets until the container is
started.

Without this change, it is not possible to use the official image
with a Dropbox or Google Drive backend.

However, this change has two negative side-effects:
1) The image size jumps from 139MB to 819MB, because of the need
   to include the node dependencies in order to build at runtime.
2) The startup time is slightly slower.