Fix webdav settings load issue
When loading the settings file, configFileContents is sometimes already
an object. Therefore, when JSON.parse is run, it throws an error which
is silently swallowed by the catch below. This appears to be the cause
of #472. Settings will never load from file and default config always
laods.

Adding a type check and assuming non strings are valid object causes the
settings to load correctly.