WIP: Links to non-org files go to GitLab
This work in progress / proof-of-concept shows a potential solution for
links to files that are not org files or folders (#325). It does this by:

- saving the URL to the GitLab project at login time
- parsing links to files that contain a "." but are not org files
- making a link to the file with the template
  `${gitLablProjectUrl}/-/blob/${defaultBranch}/${path}`

The commit is not production ready and has not added or updated any of
the tests, but it meant to solicit early feedback on this approach.
Notably it _only_ works for GitLab sync backends (though others will
revert to the default behaviour.)