fix: repeater state changes will now be filed to the correct header
I found this error when I added a headline to the end of `various_todos` fixtures, right
after the "Repeating task" header. Adding a new header caused the
'should advance repeating task' test to fail. It failed at the line where it asserted the `logNotes` of the
entry with a repeater had changed. I was able to trace this error to addNoteGeneric where I noticed
`headerId` was undefined as `state.get('selectedHeaderId')` returned
undefined. However, this would never cause `addNoteGeneric` to fail
since `headerId` is passed to `indexOfHeaderWithId`, which relies on
`findIndex`. `findIndex` returns -1 whenever it can find an element
that satisfies your testing function. That is the reason
`updatePlanningItemsWithRepeaters` appeared to work correctly. That is
also why the 'should advance repeating task' test would fail whenever
a header was added to `various_todos`.  To fix this, I added two more
headers to `various_todos`. Then, I revised 'should advance repeating
task' , 'should advance repeating task again' and 'should advance
active timestamp with repeater in header'. Finally, I made sure to select the
header that was being edited in `updatePlanningItemsWithRepeaters`.