fix: Made it harder to swipe left or right on a header accidentally
SWIPE_ACTION_ACTIVATION_DISTANCE was the root cause of the accidental
deletes. Because of that variable, you barely had to swipe at all in
order to shift the state of a header or delete it. In any case, a hard
coded variable should not have been used in this instance. Instead of using
that variable, dragStartX and currentDragX are the only variables used
to determine whether or not a state shift or delete should be
triggered. That means a state shift or delete is only triggered once a
user has swiped about halfway across the screen. The swipe animations
themselves were not touched. They still use the
SWIPE_ACTION_ACTIVATION_DISTANCE variable. So you will still see the
state shift or delete animation as soon as you began a swipe.