fix: Fixed timestamp error in OrgFile.integration.test.js
The `Clicking the Timestamp in a TODO within the agenda toggles from
the date to the time` test failed when I ran yarn test this
morning. However, the feature itself still worked when I tested it. I was able to trace
the test error to `dateForTimestamp`. `dateForTimestamp` sets the start
hour of a timestamp to 12:00 if it receives a timestamp without a start time. The expected
date in the test did not take this account. Now that I have taken that
into account, the test no longer fails.